♻️ (1396): création d'un helper construireQuery + déplacer tout dans meme sous dossier
