+++ src/modules/project/queries/index.ts
@@ -12,6 +12,4 @@ export * from './GetProjectsByContactEmail'
 export * from './GetUnnotifiedProjectsForPeriode'
 export * from './HasDemandeDeMêmeTypeOuverte'
 export * from './TrouverProjetsParIdentifiantGestionnaireRéseau'
-export * from './listProjects'
-export * from './listerProjetsPourAdmin'
-export * from './listerProjetsPourDreal'
+export * from './listerProjets'
